requirements.txt
python-dotenv
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
langchain
langchain-nvidia-ai-endpoints
#langchain-google-genai
langchain-chroma
langchain-core
langchain-community
llama-cloud-services
tenacity
google-genai
python-multipart